refactor: Question, Answer 의 기본 정보를 BaseEntity로 분리
